+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css'
 import { BrowserRouter as Router, Routes ,Route } from 'react-router-dom'
-import Register from './register_login'
-import Home from './Home'
+import Register from './pages/register_login'
+import Home from './pages/Home'
 
 
 function App() {
