+++ frontend/src/pages/register_login.tsx
@@ -1,7 +1,6 @@
 import { FormEvent, useState } from 'react';
 import axios from 'axios';
 import './App.css'
-// import { Link, Navigate } from 'react-router-dom';
 import { useNavigate } from "react-router-dom";
